requirements.txt
pandas==2.2.3
selenium==4.28.1
webdriver-manager==4.0.2
python-telegram-bot==21.10